pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-lesson-planner"
version = "2.0.0"
description = "AI-powered lesson planning application with comprehensive features"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["education", "lesson-planning", "ai", "streamlit", "teaching"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Topic :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "langchain-core>=0.3.0",
    "langchain-groq>=0.3.0",
    "streamlit>=1.39.0",
    "pymongo>=4.15.3",
    "bcrypt>=5.0.0",
    "python-dotenv>=1.1.1",
    "groq>=0.32.0",
    "python-docx>=1.1.0",
    "markdown2>=2.4.10",
    "reportlab>=4.0.7",
    "weasyprint>=60.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/ai-lesson-planner"
Documentation = "https://github.com/your-username/ai-lesson-planner/blob/main/README.md"
Repository = "https://github.com/your-username/ai-lesson-planner"
Issues = "https://github.com/your-username/ai-lesson-planner/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

